Implemented access control

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,17 @@ use DB;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -56,6 +67,7 @@ class PostsController extends Controller
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
         $post->save();
         return redirect('/posts')->with('success', 'Post created');
     }
@@ -81,6 +93,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // check if the user is authorized to edit the post
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts/')->with('error', 'Unauthorized page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -118,6 +134,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        // check if the user is authorized to edit the post
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts/')->with('error', 'Unauthorized page');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted Successfully!');
     }
